Update README.md
Sorry I have never used github so was not sure how to comment directly on the draft...and it duplicated a paragraph for some reason. Some more comments:


I know this is just a draft, but here are some general comments that you might find helful.

I would avoid the word "hype" etc - it has a very negative connotation which will not help you through the review process. Also you state that it is hype without really leading up to why it is hype. The opinion you are taking could be clearer up front in the first paragraph. What is the problem with an unclear definition of what "complex" means? How would the definition affect how people view/test trophic networks and ecosystem structure in general?

Is this going to be a TREE paper? I am not super sure what "graph complexity"/"random graphs" or Kolmogorov determinism is, and if you're aiming for TREE, your audience likely won't either. Maybe say that you will look at complexity through their mathematical definitions, which are less subjective. 

It sounds like your main argument is that "complexity" is subjective, even if we have metrics such as link density/interaction strengths etc -- and that we really need a null model of random complexity to be able to say if a system is simple or complex. Right? If that's the main argument, it is buried wayyy down and should be right up front.